src/main/java/POJO/Request/Pets.java
package POJO.Request;

import java.util.ArrayList;

public class Pets {

    private PetCategory category;
    private int id;
    private String name;
    private ArrayList<String> photoUrls;
    private ArrayList<PetTags> tags;
    private String status;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public PetCategory getCategory() {
        return category;
    }

    public void setCategory(PetCategory category) {
        this.category = category;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public ArrayList<String> getPhotoUrls() {
        return photoUrls;
    }

    public void setPhotoUrls(ArrayList<String> photoUrls) {
        this.photoUrls = photoUrls;
    }

    public ArrayList<PetTags> getTags() {
        return tags;
    }

    public void setTags(ArrayList<PetTags> tags) {
        this.tags = tags;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }



}
